requirements.txt
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.0.0
colorama==0.4.6
cryptography==39.0.0
filelock==3.9.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
snowflake-connector-python==3.0.0
snowflake-snowpark-python==0.10.0
typing-extensions==4.4.0
urllib3==1.26.14
